WIP implement support for several profiles
TODO: there still is a problem when you try to start a session in
another profile, but that has the same session name as a currently
running session. In this case webmacs just ends up remote-controlling
the running session. Fix this somehow by having the profile name be
part of the session name (at least as far as the IPC is concerned. it
would be nice if the names of the session*.json files was unchanged.)